WIP: remove list
Actually it is not necessary to store a list of predecessors. Coroutines are building up it self a list and can define a continuation with a linked list representation.

The continuation_data can be the type-eraser for this continuation. This can allow to set up "any kind" of coroutine (i.e.: promise) to the chain.

This refactor has 2 big benefits:
 - no more allocation for the list.
 - We can also reuse it to store a promise afterwards to execute. Like storing the result value, or making a receiver, etc.

TODO:
 - in case of exception I had a strange already done coroutine which was not finished (i.e. can be deleted)
 - Remove leftover code.