Sycnrhonization fix
await_ready always suspend. We should make available the suspended handle in this function in order to use the barrier and see what the other task status is to properly not suspending the current awaiter. It might worth it but right now seems a bit too much overhad and premature optimization.

Add memory synchronization to the result. When await_suspend doesn't suspend because the other thread already finished and stores the result memory is not synchronized between the two threads. Atomic_flag is used as a memory barrier,