Add lifetime management.
Calling destroy when coroutine is done. It is a tricky question.

promise's final_suspend block always. It results that done() function works properly but also that the coroutine doesn't fall off and its promise never will destroyed.

The big question is basically how to implement continueation in the coroutines. In final_suspend we could return with a co_awaitable object that makes the continuation. But in this way it is not possible to (or doesn't seems right now) to handle properly the "non_suspending" coroutine option. It's because then final_suspend sometimes would return the 'next' sometimes nothing when the original suspended coroutine is basically not suspended.

Further testing is necessary.