Add 'smart' suspend
The idea is not suspend the current coroutine if the async handle is already finished. Lifetime of the handles are guaranteed by the call chain.

For this it was necessary to make inner classes part of the __details namespace. Unfortunately with clang it leaded to compilation errors due to undefined types.
Also with clang-19 it was not supported to mark an inner class freind `friend class task<R>::awaiter_t`

Further tests are needed.