WIP No memory leak. All test passes.
Needs code cleanup

Last error was about when exception occurred. In this case not the suspended coroutine was the final coroutine and that's lifetime was managed by the sync wait.

Thus, we had to mark the coroutines in the chain whos lifetime is managed from the caller side and it should not be destroyed by the exection flow.