Add test controller and fix
The idea is that execution test cases should run with sync_wat and async_wait as well. test_controller and template test cases makes the magic.

Fix: It might happen that the top level coroutine is not the first in the chain but the last. In this case we need to release that continuation.